add basic crypto functions
Slightly borrowed from accdb.

Plain AES-CFB might not be the best choice, but one of the goals is that
if 100% of the code is bitrotted (let's say libnacl from 100 years ago
no longer compiles), it should be possible to recover the data just by
looking at the DB -- which is the reason I avoided things like Fernet in
my projecs.